refactor: create gdoc/doc.py with Doc and Tab Resource classes
Move all business logic from gdoc/__init__.py to gdoc/doc.py:
data classes, format helpers, API functions, comments, and tab
mutation helpers. Add Doc(Resource) for folder operations and
Tab(Resource) for single-tab file operations.

__init__.py re-exports symbols and keeps CLI commands (to be
moved to cli.py next). Update test patches to target doc.py.
